# perf: able to load local project faster