### What I did

crushing bottle necks 1 by 1

Avoids config processing in the `project` init and move to `path` to property to make loading project faster

### How to verify it

easiest way:

ipython

before:

```
In [1]: %time from ape import project
CPU times: user 2.93 s, sys: 580 ms, total: 3.51 s
Wall time: 2.57 s
```

after

```
%time from ape import project
CPU times: user 1.97 s, sys: 554 ms, total: 2.52 s
Wall time: 1.6 s
```

### Checklist

<!-- All PRs must complete the following checklist before being merged -->

- [ ] All changes are completed
- [ ] New test cases have been added
- [ ] Documentation has been updated